+++ app/Libraries/HtmlToImage/Contracts/Capture.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace App\Libraries\HtmlToImage\Contracts;
+
+
+interface Capture
+{
+    const SUCCESS = 0;
+    const FAIL = 1;
+
+    public function capture();
+}